+++ src/cli.ts
@@ -39,6 +39,7 @@ import { memoryCommand } from "./commands/memory.js";
 import { changelogCommand } from "./commands/changelog.js";
 import { a11yLintCommand } from "./commands/a11y-lint.js";
 import { flowCommand } from "./commands/flow.js";
+import { contentLintCommand } from "./commands/content-lint.js";
 import { ingestFigmaDsCommand } from "./commands/ingest-figma-ds.js";
 import { synthesizeConventionsCommand } from "./commands/synthesize-conventions.js";
 
@@ -81,6 +82,7 @@ COMMANDS[memoryCommand.name] = memoryCommand;
 COMMANDS[changelogCommand.name] = changelogCommand;
 COMMANDS[a11yLintCommand.name] = a11yLintCommand;
 COMMANDS[flowCommand.name] = flowCommand;
+COMMANDS[contentLintCommand.name] = contentLintCommand;
 COMMANDS[ingestFigmaDsCommand.name] = ingestFigmaDsCommand;
 COMMANDS[synthesizeConventionsCommand.name] = synthesizeConventionsCommand;